requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_py-xgboost-mutex=2.0=cpu_0
absl-py=1.2.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
atk-1.0=2.36.0=ha1a6a79_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pyhd3eb1b0_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py39hce1f21e_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39hb9d737c_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
cairo=1.16.0=h19f5f5c_2
certifi=2022.6.15=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.1.0=pyhd8ed1ab_0
cryptography=37.0.2=py39hd97740a_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dgl-cuda11.6=0.9.0=py39_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.13.1=h6c09931_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fribidi=1.0.10=h7b6447c_0
frozenlist=1.3.1=pypi_0
fsspec=2022.7.1=pypi_0
gdk-pixbuf=2.42.8=h433bba3_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gobject-introspection=1.68.0=py39he41a700_3
google-auth=2.11.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h3cd0ef9_0
grpcio=1.47.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
h5py=3.7.0=py39h737f45e_0
harfbuzz=4.3.0=hd55b92a_0
hdf5=1.10.6=hb1b8bf9_0
hydra-core=1.2.0=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py39h06a4308_0
ipympl=0.9.1=pyhd8ed1ab_0
ipython=8.3.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonschema=4.4.0=py39h06a4308_0
jupyter=1.0.0=py39h06a4308_7
jupyter_client=7.2.2=py39h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py39h295c915_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgd=2.3.3=h695aa2c_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libllvm11=11.1.0=h3826bc1_1
libpng=1.6.37=hbc83047_0
librsvg=2.54.4=h19fe530_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h2818925_1
libtool=2.4.6=h295c915_1008
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxgboost=1.5.0=h6a678d5_2
libxml2=2.9.14=h74e7548_0
littleutils=0.2.2=py_0
llvmlite=0.38.0=py39h4ff587b_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.5.1=py39h06a4308_1
matplotlib-base=3.5.1=py39ha18d171_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=pypi_0
munkres=1.1.4=py_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.3.0=py39h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.6=he412f7d_0
networkx=2.8.5=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
notebook=6.4.11=py39h06a4308_0
numba=0.55.1=py39h51133e4_0
numexpr=2.8.1=py39h807cd23_2
numpy=1.21.5=py39h6c91a56_3
numpy-base=1.21.5=py39ha15fc14_3
oauthlib=3.2.0=pypi_0
ogb=1.3.3=pyhd8ed1ab_0
omegaconf=2.2.3=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1q=h7f8727e_0
outdated=0.2.1=pyhd8ed1ab_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39h295c915_0
pandocfilters=1.5.0=pyhd3eb1b0_0
pango=1.50.7=h05da053_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.4=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-xgboost=1.5.0=py39h06a4308_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39hf3d152e_5
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-lightning=1.7.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py39h06a4308_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py39h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
root=6.26.6=py38h2c36dcf_0
root_base=6.26.6=py38h259b632_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.9.0=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.5=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.5.0=hd09550d_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.1=py39h06a4308_0
testpath=0.6.0=py39h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torch-cluster=1.6.0=pypi_0
torch-geometric=2.1.0=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.15=pypi_0
torchaudio=0.12.1=py39_cu116
torchmetrics=0.9.3=pyhd8ed1ab_0
torchvision=0.13.1=py39_cu116
tornado=6.1=py39h27cfd23_0
tqdm=4.64.0=py39h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.11=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h06a4308_0
xrootd=5.4.3=py38h92bdfd5_1
xz=5.2.5=h7f8727e_1
yarl=1.8.1=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
